+++ src/filecont.h
@@ -46,8 +46,8 @@ public:
     virtual bool DeleteFile(const std::string &wildcard) = 0;
     virtual bool RenameFile(const std::string &oldName,
                             const std::string &newName) = 0;
-    virtual bool SetAttributes(const std::string &wildcard, Byte setMask,
-                               Byte clearMask = ~0) = 0;
+    virtual bool SetAttributes(const std::string &wildcard, unsigned setMask,
+                               unsigned clearMask = ~0U) = 0;
     virtual FlexFileBuffer ReadToBuffer(const std::string &fileName) = 0;
     virtual bool WriteFromBuffer(const FlexFileBuffer &buffer,
                                  const char *fileName = nullptr) = 0;
